+++ exporter/tanzuobservabilityexporter/metrics.go
@@ -18,6 +18,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"math"
 	"strconv"
 	"sync/atomic"
 
@@ -50,6 +51,9 @@ var (
 var (
 	// Specifies regular histogram
 	regularHistogram histogramConsumerSpec = regularHistogramConsumerSpec{}
+
+	// Specifies exponential histograms
+	exponentialHistogram histogramConsumerSpec = exponentialHistogramConsumerSpec{}
 )
 
 // metricsConsumer instances consume OTEL metrics
@@ -608,5 +612,219 @@ func (regularHistogramConsumerSpec) AsHistogram(metric pdata.Metric) histogramMe
 	return &regularHistogramMetric{
 		Histogram:               aHistogram,
 		HistogramDataPointSlice: aHistogram.DataPoints(),
+	}
+}
+
+type summaryConsumer struct {
+	sender   gaugeSender
+	settings component.TelemetrySettings
+}
+
+// newSummaryConsumer returns a typedMetricConsumer that consumes summary metrics
+// by sending them to tanzu observability.
+func newSummaryConsumer(
+	sender gaugeSender, settings component.TelemetrySettings,
+) typedMetricConsumer {
+	return &summaryConsumer{sender: sender, settings: settings}
+}
+
+func (s *summaryConsumer) Type() pdata.MetricDataType {
+	return pdata.MetricDataTypeSummary
+}
+
+func (s *summaryConsumer) Consume(metric pdata.Metric, errs *[]error) {
+	summary := metric.Summary()
+	summaryDataPoints := summary.DataPoints()
+	for i := 0; i < summaryDataPoints.Len(); i++ {
+		s.sendSummaryDataPoint(metric, summaryDataPoints.At(i), errs)
+	}
+}
+
+// PushInternalMetrics is here so that summaryConsumer implements typedMetricConsumer
+func (*summaryConsumer) PushInternalMetrics(*[]error) {
+	// Do nothing
+}
+
+func (s *summaryConsumer) sendSummaryDataPoint(
+	metric pdata.Metric, summaryDataPoint pdata.SummaryDataPoint, errs *[]error,
+) {
+	name := metric.Name()
+	ts := summaryDataPoint.Timestamp().AsTime().Unix()
+	tags := attributesToTags(summaryDataPoint.Attributes())
+	count := summaryDataPoint.Count()
+	sum := summaryDataPoint.Sum()
+
+	if quantileTag, ok := tags["quantile"]; ok {
+		tags["_quantile"] = quantileTag
+		delete(tags, "quantile")
+	}
+	s.sendMetric(name+"_count", float64(count), ts, tags, errs)
+	s.sendMetric(name+"_sum", sum, ts, tags, errs)
+	quantileValues := summaryDataPoint.QuantileValues()
+	for i := 0; i < quantileValues.Len(); i++ {
+		quantileValue := quantileValues.At(i)
+		tags["quantile"] = quantileTagValue(quantileValue.Quantile())
+		s.sendMetric(name, quantileValue.Value(), ts, tags, errs)
+	}
+}
+
+func (s *summaryConsumer) sendMetric(
+	name string, value float64, ts int64, tags map[string]string, errs *[]error) {
+	err := s.sender.SendMetric(name, value, ts, "", tags)
+	if err != nil {
+		*errs = append(*errs, err)
+	}
+}
+
+func quantileTagValue(quantile float64) string {
+	return strconv.FormatFloat(quantile, 'f', -1, 64)
+}
+
+type exponentialHistogramDataPoint struct {
+	pdata.ExponentialHistogramDataPoint
+	bucketCounts   []uint64
+	explicitBounds []float64
+}
+
+// newExponentialHistogram converts a pdata.ExponentialHistogramDataPoint into a histogramDataPoint
+// implementation. A regular histogramDataPoint has bucket counts and explicit bounds for each
+// bucket; an ExponentialHistogramDataPoint has only bucket counts because the explicit bounds
+// for each bucket are implied because they grow exponentially from bucket to bucket. The
+// conversion of an ExponentialHistogramDataPoint to a histogramDataPoint is necessary because the
+// code that sends histograms to tanzuobservability expects the histogramDataPoint format.
+func newExponentialHistogramDataPoint(dataPoint pdata.ExponentialHistogramDataPoint) histogramDataPoint {
+
+	// Base is the factor by which the explicit bounds increase from bucket to bucket.
+	// This formula comes from the documentation here:
+	// https://github.com/open-telemetry/opentelemetry-proto/blob/8ba33cceb4a6704af68a4022d17868a7ac1d94f4/opentelemetry/proto/metrics/v1/metrics.proto#L487
+	base := math.Pow(2.0, math.Pow(2.0, -float64(dataPoint.Scale())))
+
+	// ExponentialHistogramDataPoints have buckets with negative explicit bounds, buckets with
+	// positive explicit bounds, and a "zero" bucket. Our job is to merge these bucket groups into
+	// a single list of buckets and explicit bounds.
+	negativeBucketCounts := dataPoint.Negative().BucketCounts()
+	positiveBucketCounts := dataPoint.Positive().BucketCounts()
+
+	// The total number of buckets is the number of negative buckets + the number of positive
+	// buckets + 1 for the zero bucket + 1 bucket for negative infinity up to the negative explicit
+	// bound with largest magnitude + 1 bucket for the largest positive explicit bound up to
+	// positive infinity.
+	numBucketCounts := 1 + len(negativeBucketCounts) + 1 + len(positiveBucketCounts) + 1
+
+	// We pre-allocate the slice setting its length to 0 so that GO doesn't have to keep
+	// re-allocating the slice as it grows.
+	bucketCounts := make([]uint64, 0, numBucketCounts)
+
+	// The number of explicit bounds is always 1 less than the number of buckets. This is how
+	// explicit bounds work. If you have 2 explicit bounds say {2.0, 5.0} then you have 3 buckets:
+	// one for values less than 2.0; one for values between 2.0 and 5.0; and one for values greater
+	// than 5.0.
+	explicitBounds := make([]float64, 0, numBucketCounts-1)
+
+	appendNegativeBucketsAndExplicitBounds(
+		dataPoint.Negative().Offset(), base, negativeBucketCounts, &bucketCounts, &explicitBounds)
+	appendZeroBucketAndExplicitBound(
+		dataPoint.Positive().Offset(), base, dataPoint.ZeroCount(), &bucketCounts, &explicitBounds)
+	appendPositiveBucketsAndExplicitBounds(
+		dataPoint.Positive().Offset(), base, positiveBucketCounts, &bucketCounts, &explicitBounds)
+	return &exponentialHistogramDataPoint{
+		ExponentialHistogramDataPoint: dataPoint,
+		bucketCounts:                  bucketCounts,
+		explicitBounds:                explicitBounds,
+	}
+}
+
+// appendNegativeBucketsAndExplicitBounds appends negative buckets and explicit bounds to
+// bucketCounts and explicitBounds respectively. The largest negative explicit bound (the one
+// with the smallest magnitude) is -1*base^negativeOffset
+func appendNegativeBucketsAndExplicitBounds(
+	negativeOffset int32,
+	base float64,
+	negativeBucketCounts []uint64,
+	bucketCounts *[]uint64,
+	explicitBounds *[]float64,
+) {
+	// The count in the first bucket which includes negative infinity is always 0.
+	*bucketCounts = append(*bucketCounts, 0)
+
+	// The smallest negative explicit bound.
+	le := -math.Pow(base, float64(negativeOffset)+float64(len(negativeBucketCounts)))
+	*explicitBounds = append(*explicitBounds, le)
+
+	// The first negativeBucketCount has a negative explicit bound with the smallest magnitude;
+	// the last negativeBucketCount has a negative explicit bound with the largest magnitude.
+	// Therefore, to go in order from smallest to largest explicit bound, we have to start with
+	// the last element in the negativeBucketCounts array.
+	for i := len(negativeBucketCounts) - 1; i >= 0; i-- {
+		*bucketCounts = append(*bucketCounts, negativeBucketCounts[i])
+		le /= base // We divide by base because our explicit bounds are getting larger as we go
+		*explicitBounds = append(*explicitBounds, le)
+	}
+}
+
+// appendZeroBucketAndExplicitBound appends the "zero" bucket and explicit bound to bucketCounts
+// and explicitBounds respectively. The smallest positive explicit bound is base^positiveOffset.
+func appendZeroBucketAndExplicitBound(
+	positiveOffset int32,
+	base float64,
+	zeroBucketCount uint64,
+	bucketCounts *[]uint64,
+	explicitBounds *[]float64,
+) {
+	*bucketCounts = append(*bucketCounts, zeroBucketCount)
+
+	// The explicit bound of the zeroBucketCount is the smallest positive explicit bound
+	*explicitBounds = append(*explicitBounds, math.Pow(base, float64(positiveOffset)))
+}
+
+// appendPositiveBucketsAndExplicitBounds appends positive buckets and explicit bounds to
+// bucketCounts and explicitBounds respectively. The smallest positive explicit bound is
+// base^positiveOffset.
+func appendPositiveBucketsAndExplicitBounds(
+	positiveOffset int32,
+	base float64,
+	positiveBucketCounts []uint64,
+	bucketCounts *[]uint64,
+	explicitBounds *[]float64,
+) {
+	le := math.Pow(base, float64(positiveOffset))
+	for _, bucketCount := range positiveBucketCounts {
+		*bucketCounts = append(*bucketCounts, bucketCount)
+		le *= base
+		*explicitBounds = append(*explicitBounds, le)
+	}
+	// Last bucket count for positive infinity is always 0.
+	*bucketCounts = append(*bucketCounts, 0)
+}
+
+func (e *exponentialHistogramDataPoint) ExplicitBounds() []float64 {
+	return e.explicitBounds
+}
+
+func (e *exponentialHistogramDataPoint) BucketCounts() []uint64 {
+	return e.bucketCounts
+}
+
+type exponentialHistogramMetric struct {
+	pdata.ExponentialHistogram
+	pdata.ExponentialHistogramDataPointSlice
+}
+
+func (e *exponentialHistogramMetric) At(i int) histogramDataPoint {
+	return newExponentialHistogramDataPoint(e.ExponentialHistogramDataPointSlice.At(i))
+}
+
+type exponentialHistogramConsumerSpec struct {
+}
+
+func (exponentialHistogramConsumerSpec) Type() pdata.MetricDataType {
+	return pdata.MetricDataTypeExponentialHistogram
+}
+
+func (exponentialHistogramConsumerSpec) AsHistogram(metric pdata.Metric) histogramMetric {
+	aHistogram := metric.ExponentialHistogram()
+	return &exponentialHistogramMetric{
+		ExponentialHistogram:               aHistogram,
+		ExponentialHistogramDataPointSlice: aHistogram.DataPoints(),
 	}
 }